Adapt base and plots to run regions of countries
This commit adds support for the forecasting and plotting of subregions
for which case and death data is available and mapping all other "state"
information to the country of which it is part of. This is good for the
uniform application of NPIs but is certainly an issue with regard to
age distributions.