Functionalise covariate reading and processing
Moves covariates reading and preprocessing to a function. This enables
both base.r and plot-3-panel to use the same process.

This avoids copy pasted code, and makes sure that what is plotted is
what is calculated. This commit changes the 3 panel plots for spain
now showing self isolation applied at the same time as lockdown rather
than after.